refactor(ReceiptModel): update and seperate ReceiptModel from ReceiptText
